requirements.txt
Django==1.5.1
chardet==2.1.1
django-adaptors==0.2.5
django-appconf==0.6
django-csvimport==1.0
django-geoposition==0.1.4
django-grappelli==2.4.5
django-money==0.3.3
django-taggit==0.10a1
geopy==0.95.1
lxml==3.2.1
py==1.4.15
py-moneyed==0.4.1
pytest==2.3.5
six==1.3.0
unicodecsv==0.9.4
wsgiref==0.1.2
